Refactor to_backward()
- Put it in a dedicated module decomon.backward_layers.convert
- Instead of globals() use a mapping constructed from vars() applied to
  relevant module. It avoids explicitely import each BackwardLayer and
  make the module imports be used so that it is obvious even for IDE that
  such imports are not to be removed.
