+++ src/decomon/backward_layers/__init__.py
@@ -1,1 +1,1 @@
-from .backward_layers import to_backward
+from .convert import to_backward
