+++ tests/test_backward_dense.py
@@ -7,7 +7,7 @@ import tensorflow.keras.backend as K
 from tensorflow.keras.layers import Dense, Input
 from tensorflow.keras.models import Model
 
-from decomon.backward_layers.backward_layers import to_backward
+from decomon.backward_layers.convert import to_backward
 from decomon.layers.core import ForwardMode
 from decomon.layers.decomon_layers import DecomonDense
 from decomon.utils import Slope
