+++ tests/test_decomon_activation_layer.py
@@ -1,13 +1,7 @@
-import numpy as np
-import pytest
 import tensorflow.python.keras.backend as K
-from tensorflow.keras.layers import Input
-from tensorflow.keras.models import Model
 
-from decomon.backward_layers.backward_layers import to_backward
 from decomon.layers.core import ForwardMode
-from decomon.layers.decomon_layers import DecomonActivation, DecomonFlatten
-from decomon.layers.decomon_reshape import DecomonReshape
+from decomon.layers.decomon_layers import DecomonActivation
 from decomon.utils import Slope
